Support for `DOUBLE_CONST`, update esprima with improved literal parsing

+++ GTAdhocToolchain.Core/Instructions/InsDoubleConst.cs
@@ -14,7 +14,7 @@ namespace GTAdhocToolchain.Core.Instructions
 
         public double Value { get; set; }
 
-        public InsDoubleConst(ulong value)
+        public InsDoubleConst(double value)
         {
             Value = value;
         }
